% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Plots counts of sequence feature}
\usage{
plot_counts(physeq, otu, group)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{otu}{Name of the otu/asv to be visualized}

\item{group}{Grouping variable to stratify results by}
}
\value{
An object of type ggplot2
}
\description{
Plots counts of sequence feature
}
\examples{
physeq <- readRDS(system.file("extdata", "MUClite.rds", package="decontam"))
plot_counts(physeq, otu = "Seq1", group = 'Habitat')
}
