% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_statistics.R
\name{plot_score_statistics}
\alias{plot_score_statistics}
\title{Plot score statistics output from decontam}
\usage{
plot_score_statistics(x)
}
\arguments{
\item{x}{A data.frame of classification information output from the isContaminant function in the decontam package.}
}
\value{
A ggplot object showing the distribution of score statistics assigned to each sequence feature.
}
\description{
A visualization of the distribution of score statistics output from the decontam package (Davis et al., 2018).  As the program authors note, a visualization of the score statistics is needed to inform a threshold for which to determine if a sequence feature is a contaminant or not.
}
\examples{
physeq <- readRDS(system.file("extdata", "MUClite.rds", package="decontam"))
contam_results <- decontam::isContaminant(physeq, method = 'frequency', conc="quant_reading")
plot_score_statistics(contam_results)
}
