% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_contaminants.R
\name{prune_contaminants}
\alias{prune_contaminants}
\title{Prune contaminants from a phyloseq object}
\usage{
prune_contaminants(x, physeq)
}
\arguments{
\item{x}{A data.frame of classification information output from the isContaminant function in the decontam package.}

\item{physeq}{A phyloseq object}
}
\value{
A trimmed phyloseq object with contaminants removed
}
\description{
Prune contaminants from a phyloseq object
}
\examples{
physeq <- readRDS(system.file("extdata", "MUClite.rds", package="decontam"))
contam_results <- decontam::isContaminant(physeq, method = 'frequency', conc="quant_reading")
physeq.filt <- prune_contaminants(contam_results, physeq)
}
